pyproject.toml
[tool.poetry]
name = "django-drf-hw"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.6"
djangorestframework = "^3.15.2"
pillow = "^11.1.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
django-filter = "^25.1"
djangorestframework-simplejwt = "^5.5.0"
pytest-cov = "^6.0.0"
coverage = "^7.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
